added types to useCustomCookies hook
